requirements.txt
bs4
googlesearch
requests
requests_random_user_agent
random
lxml.html
itertools
traceback
time